# TelemetryWerk.Ui/TelemetryWerk.Ui.Infrastructure/OpenAPI/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "TelemetryWerk.Api.Host",
    "version": "1.0"
  },
  "paths": {
    "/api/v1/Telemetry/nodes": {
      "get": {
        "tags": [
          "Telemetry"
        ],
        "parameters": [
          {
            "name": "afterId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MachineNodeDtoPagedCollectionUnifiedResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineNodeDtoPagedCollectionUnifiedResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineNodeDtoPagedCollectionUnifiedResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Telemetry"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MachineNodeDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/MachineNodeDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/MachineNodeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MachineNodeDtoUnifiedResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineNodeDtoUnifiedResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineNodeDtoUnifiedResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/Telemetry/nodes/{id}": {
      "patch": {
        "tags": [
          "Telemetry"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNodeRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNodeRequestDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNodeRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MachineNodeDtoUnifiedResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineNodeDtoUnifiedResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineNodeDtoUnifiedResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Telemetry"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "MachineNodeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "coreTemperature": {
            "type": "number",
            "format": "double"
          },
          "status": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "MachineNodeDtoPagedCollection": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MachineNodeDto"
            },
            "nullable": true
          },
          "lastId": {
            "type": "string",
            "nullable": true
          },
          "hasMore": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "MachineNodeDtoPagedCollectionUnifiedResponse": {
        "type": "object",
        "properties": {
          "status": {
            "$ref": "#/components/schemas/StatusResponseDto"
          },
          "data": {
            "$ref": "#/components/schemas/MachineNodeDtoPagedCollection"
          }
        },
        "additionalProperties": false
      },
      "MachineNodeDtoUnifiedResponse": {
        "type": "object",
        "properties": {
          "status": {
            "$ref": "#/components/schemas/StatusResponseDto"
          },
          "data": {
            "$ref": "#/components/schemas/MachineNodeDto"
          }
        },
        "additionalProperties": false
      },
      "StatusResponseDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "serverTime": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UpdateNodeRequestDto": {
        "type": "object",
        "properties": {
          "coreTemperature": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "ApiKey": {
        "type": "apiKey",
        "description": "API Key authentication. Enter your API Key below.\nExample: DEV_MODE_PLEASE_CHANGE_THIS_KEY_IN_PRODUCTION",
        "name": "X-Api-Key",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "ApiKey": [ ]
    }
  ]
}